airflow/requirements.txt
spotipy 
pandas 
matplotlib 
seaborn

apache-airflow-providers-google>=5.0.0
apache-airflow[google]
apache-airflow-providers-apache-spark
pyarrow
py4j
pyspark
pandas
google-cloud-storage